Backend/requirements.txt
sqlmodel==0.0.22
uvicorn==0.34.0
fastapi==0.115.6
asyncmy==0.2.10
python-dotenv==1.0.1
